strlen

+++ pipex_alt/src/bonus/string_bonus.c
@@ -6,15 +6,15 @@
 /*   By: athirion <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/24 14:56:36 by athirion          #+#    #+#             */
-/*   Updated: 2022/02/24 14:58:31 by athirion         ###   ########.fr       */
+/*   Updated: 2022/03/01 18:40:30 by athirion         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "pipex.h"
+#include "pipex_bonus.h"
 
-size_t	ft_strlen(const char *s)
+int	ft_strlen(const char *s)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (s[i])
